decouple file path operation logic
